atualizando os fontes - Apresentação 2 socialização

+++ app/src/main/java/com/yuri/luis/garcia/pereira/tcs_implementacao/model/Execucao.kt
@@ -9,7 +9,7 @@ data class Execucao(
     val regras: List<ExecucaoRegra>,
     val image: Image?,
     val concluido: String?,
-    val percentualAcerto :Int?
+    val percentualAcerto :Double?
 ) {
     constructor(): this(null,
         null, null, emptyList(), null,null,null)
